refactor: move authoring LSP into a cell
