+++ cells/cell-authoring-lsp-proto/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cell-authoring-lsp-proto"
+version = "0.6.1"
+edition = "2024"
+rust-version = "1.91"
+
+[package.metadata.docs.rs]
+rustdoc-args = ["--html-in-header", "arborium-header.html"]
+
+[package.metadata."docs.rs"]
+rustdoc-args = ["--html-in-header", "arborium-header.html"]
+
+[dependencies]
+facet.workspace = true
+vox.workspace = true
