+++ crates/dodeca/Cargo.toml
@@ -24,6 +24,7 @@ path = "src/main.rs"
 
 [dependencies]
 cell-code-execution-proto = { path = "../../cells/cell-code-execution-proto" }
+cell-authoring-lsp-proto = { path = "../../cells/cell-authoring-lsp-proto" }
 cell-vite-proto = { path = "../../cells/cell-vite-proto" }
 dodeca-config = { path = "../dodeca-config" }
 cell-lifecycle-proto = { path = "../../cells/cell-lifecycle-proto" }
@@ -63,7 +64,6 @@ figue.workspace = true
 facet-json.workspace = true
 facet-postcard.workspace = true
 facet-styx.workspace = true
-facet-toml.workspace = true
 styx-embed.workspace = true
 styx-tree.workspace = true
 facet-value.workspace = true
@@ -77,7 +77,6 @@ notify.workspace = true
 open.workspace = true
 owo-colors.workspace = true
 picante.workspace = true
-pulldown-cmark.workspace = true
 vox = { workspace = true, features = ["transport-all", "transport-ffi"] }
 vox-ffi.workspace = true
 libloading.workspace = true
@@ -87,7 +86,6 @@ rapidhash.workspace = true
 arboard = { version = "3", default-features = false }
 strid.workspace = true
 tokio.workspace = true
-tower-lsp.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 ur-taking-me-with-you.workspace = true
@@ -95,7 +93,6 @@ hotmeal.workspace = true
 hotmeal-server.workspace = true
 html-escape.workspace = true
 ariadne.workspace = true
-serde_json.workspace = true
 
 [build-dependencies]
 dodeca-config = { path = "../dodeca-config" }
