+++ crates/dodeca/src/lib.rs
@@ -1,0 +1,47 @@
+#![allow(clippy::collapsible_if)]
+
+// Embed Styx schema for tooling discovery (generated by build.rs).
+styx_embed::embed_outdir_file!("schema.styx");
+
+pub mod authoring_model;
+pub mod boot_state;
+pub mod build_context;
+pub mod build_steps;
+pub mod cache_bust;
+pub mod cas;
+pub mod cell_loader;
+pub mod cell_server;
+pub mod cells;
+pub mod config;
+pub mod content_service;
+pub mod data;
+pub mod db;
+pub mod error_pages;
+pub mod file_watcher;
+pub mod frontmatter_schema;
+pub mod host;
+pub mod image;
+pub mod init;
+pub mod link_checker;
+pub mod logging;
+pub mod queries;
+pub mod render;
+pub mod revision;
+pub mod search;
+pub mod serve;
+pub mod spawn;
+pub mod svg;
+pub mod template_host;
+pub mod theme_resolver;
+pub mod tui;
+pub mod tui_host;
+pub mod types;
+pub mod url_rewrite;
+pub mod vite;
+
+pub use build_context::{BuildContext, is_data_file_extension};
+
+pub fn dodeca_version() -> &'static str {
+    let version = option_env!("DODECA_RELEASE_VERSION").unwrap_or(env!("CARGO_PKG_VERSION"));
+    version.strip_prefix('v').unwrap_or(version)
+}
