+++ crates/dodeca/src/main.rs
@@ -1,76 +1,30 @@
 #![allow(clippy::collapsible_if)]
 
-// Embed Styx schema for tooling discovery (generated by build.rs)
-styx_embed::embed_outdir_file!("schema.styx");
-
-mod authoring_lsp;
-mod authoring_model;
-mod boot_state;
-mod build_steps;
-mod cache_bust;
-mod cas;
-mod cell_loader;
-mod cell_server;
-mod cells;
-mod config;
-mod content_service;
-mod data;
-mod db;
-mod error_pages;
-mod file_watcher;
-mod frontmatter_schema;
-mod host;
-mod image;
-mod init;
-mod link_checker;
-mod logging;
-mod queries;
-mod render;
-mod revision;
-mod search;
-mod serve;
-mod spawn;
-mod svg;
-mod template_host;
-mod theme_resolver;
-mod tui;
-mod tui_host;
-mod types;
-mod url_rewrite;
-mod vite;
-
-use crate::config::{LinkCheckMode, ResolvedConfig};
-use crate::db::{
-    DataFile, DataRegistry, Database, MarkdownRenderSettings, OutputFile, QueryStats, SassFile,
-    SassRegistry, SourceFile, SourceRegistry, StaticFile, StaticRegistry, TemplateFile,
+use camino::{Utf8Path, Utf8PathBuf};
+use dodeca::config::{LinkCheckMode, ResolvedConfig};
+use dodeca::db::{
+    self, DataFile, DataRegistry, Database, MarkdownRenderSettings, OutputFile, QueryStats,
+    SassFile, SassRegistry, SourceFile, SourceRegistry, StaticFile, StaticRegistry, TemplateFile,
     TemplateRegistry,
 };
-use crate::queries::build_site;
-use crate::tui::LogEvent;
-use crate::types::{
-    DataContent, DataPath, Route, SassContent, SassPath, SassPathRef, SourceContent, SourcePath,
-    SourcePathRef, StaticPath, TemplateContent, TemplatePath, TemplatePathRef,
+use dodeca::queries::{self, build_site};
+use dodeca::tui::{self, LogEvent};
+use dodeca::types::{
+    DataContent, DataPath, Route, SassContent, SassPath, SourceContent, SourcePath, StaticPath,
+    TemplateContent, TemplatePath,
 };
-use camino::{Utf8Path, Utf8PathBuf};
+use dodeca::{
+    BuildContext, cas, cell_server, cells, file_watcher, host, init, is_data_file_extension,
+    link_checker, logging, render, serve, tui_host, vite,
+};
 use eyre::{Result, eyre};
 use facet::Facet;
 use figue::{self as args, FigueBuiltins};
 use ignore::WalkBuilder;
 use owo_colors::OwoColorize;
-use std::collections::{BTreeMap, HashSet};
+use std::collections::HashSet;
 use std::fs;
 use std::sync::Arc;
-
-fn dodeca_version() -> &'static str {
-    let version = option_env!("DODECA_RELEASE_VERSION").unwrap_or(env!("CARGO_PKG_VERSION"));
-    version.strip_prefix('v').unwrap_or(version)
-}
-
-/// Check if a file extension is a supported data file format
-fn is_data_file_extension(ext: &str) -> bool {
-    let ext_lower = ext.to_lowercase();
-    matches!(ext_lower.as_str(), "json" | "toml" | "yaml" | "yml")
-}
 
 /// ddc - Static site generator
 #[derive(Facet, Debug)]
@@ -310,7 +264,7 @@ fn resolve_dirs(
     match config {
         Some(cfg) => {
             // Initialize global config for access from render pipeline
-            crate::config::set_global_config(cfg.clone())?;
+            dodeca::config::set_global_config(cfg.clone())?;
             let content_dir = content.unwrap_or(cfg.content_dir);
             let output_dir = output.unwrap_or(cfg.output_dir);
             Ok(ResolvedBuildConfig {
@@ -347,7 +301,7 @@ fn main() -> Result<()> {
         .expect("failed to build args schema")
         .cli(|cli| cli.args(cli_args))
         .env(|env| env.prefix("DDC"))
-        .help(|h| h.program_name("ddc").version(dodeca_version()))
+        .help(|h| h.program_name("ddc").version(dodeca::dodeca_version()))
         .build();
 
     let args = args::Driver::new(config).run().unwrap();
@@ -437,7 +391,7 @@ async fn async_main(command: Command) -> Result<()> {
             }
         }
         Command::Clean(args) => {
-            use crate::config::ProjectPaths;
+            use dodeca::config::ProjectPaths;
 
             // Get project paths from config
             let paths = if let Some(p) = args.path {
@@ -513,7 +467,22 @@ async fn async_main(command: Command) -> Result<()> {
         }
         Command::Lsp(args) => {
             logging::init_standard_tracing();
-            authoring_lsp::run(args.content, args.output).await
+            let client = cells::authoring_lsp_cell()
+                .await
+                .ok_or_else(|| eyre!("Authoring LSP cell not available"))?;
+            match client
+                .run(cell_authoring_lsp_proto::AuthoringLspStartupArgs {
+                    content: args.content,
+                    output: args.output,
+                })
+                .await
+            {
+                Ok(cell_authoring_lsp_proto::AuthoringLspRunResult::Success) => Ok(()),
+                Ok(cell_authoring_lsp_proto::AuthoringLspRunResult::Error { message }) => {
+                    Err(eyre!(message))
+                }
+                Err(error) => Err(eyre!("Authoring LSP cell failed: {error}")),
+            }
         }
         Command::Static(args) => {
             let path = Utf8PathBuf::from(&args.path);
@@ -674,412 +643,6 @@ impl Default for BuildOptions {
             progress: None,
             link_check: LinkCheckOptions::None,
         }
-    }
-}
-
-/// The build context with picante database
-pub struct BuildContext {
-    pub db: Arc<Database>,
-    pub content_dir: Utf8PathBuf,
-    pub output_dir: Utf8PathBuf,
-    /// Source files keyed by source path
-    pub sources: BTreeMap<SourcePath, SourceFile>,
-    /// Template files keyed by template path
-    pub templates: BTreeMap<TemplatePath, TemplateFile>,
-    /// Sass/SCSS files keyed by sass path
-    pub sass_files: BTreeMap<SassPath, SassFile>,
-    /// Static files keyed by static path
-    pub static_files: BTreeMap<StaticPath, StaticFile>,
-    /// Data files keyed by data path
-    pub data_files: BTreeMap<DataPath, DataFile>,
-    /// Query statistics (if tracking enabled)
-    pub stats: Option<Arc<QueryStats>>,
-}
-
-impl BuildContext {
-    pub fn new(content_dir: &Utf8Path, output_dir: &Utf8Path) -> Self {
-        Self::with_stats(content_dir, output_dir, None)
-    }
-
-    pub fn with_stats(
-        content_dir: &Utf8Path,
-        output_dir: &Utf8Path,
-        stats: Option<Arc<QueryStats>>,
-    ) -> Self {
-        let db = Arc::new(Database::new(stats.clone()));
-        Self {
-            db,
-            content_dir: content_dir.to_owned(),
-            output_dir: output_dir.to_owned(),
-            sources: BTreeMap::new(),
-            templates: BTreeMap::new(),
-            sass_files: BTreeMap::new(),
-            static_files: BTreeMap::new(),
-            data_files: BTreeMap::new(),
-            stats,
-        }
-    }
-
-    /// Get the database Arc for sharing with render contexts
-    pub fn db_arc(&self) -> Arc<Database> {
-        self.db.clone()
-    }
-
-    /// Get the templates directory (sibling to content dir)
-    pub fn templates_dir(&self) -> Utf8PathBuf {
-        self.content_dir
-            .parent()
-            .unwrap_or(&self.content_dir)
-            .join("templates")
-    }
-
-    /// Get the sass directory (sibling to content dir)
-    pub fn sass_dir(&self) -> Utf8PathBuf {
-        self.content_dir
-            .parent()
-            .unwrap_or(&self.content_dir)
-            .join("sass")
-    }
-
-    /// Get the static directory (sibling to content dir)
-    pub fn static_dir(&self) -> Utf8PathBuf {
-        self.content_dir
-            .parent()
-            .unwrap_or(&self.content_dir)
-            .join("static")
-    }
-
-    /// Get the dist directory (sibling to content dir, for generated/build output)
-    pub fn dist_dir(&self) -> Utf8PathBuf {
-        self.content_dir
-            .parent()
-            .unwrap_or(&self.content_dir)
-            .join("dist")
-    }
-
-    /// Get the data directory (sibling to content dir)
-    pub fn data_dir(&self) -> Utf8PathBuf {
-        self.content_dir
-            .parent()
-            .unwrap_or(&self.content_dir)
-            .join("data")
-    }
-
-    /// Load all source files into the database
-    pub fn load_sources(&mut self) -> Result<()> {
-        let md_files: Vec<Utf8PathBuf> = WalkBuilder::new(&self.content_dir)
-            .build()
-            .filter_map(|e| e.ok())
-            .filter(|e| e.file_type().map(|ft| ft.is_file()).unwrap_or(false))
-            .filter(|e| e.path().extension().map(|ext| ext == "md").unwrap_or(false))
-            .filter_map(|e| Utf8PathBuf::from_path_buf(e.into_path()).ok())
-            .collect();
-
-        for path in md_files {
-            let content = fs::read_to_string(&path)?;
-            let last_modified = fs::metadata(&path)?
-                .modified()?
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_secs() as i64)
-                .unwrap_or(0);
-            let relative = path
-                .strip_prefix(&self.content_dir)
-                .map(|p| p.to_string())
-                .unwrap_or_else(|_| path.to_string());
-
-            let source_path = SourcePath::new(relative);
-            let source_content = SourceContent::new(content);
-            let source = SourceFile::new(
-                &*self.db,
-                source_path.clone(),
-                source_content,
-                last_modified,
-            )?;
-            self.sources.insert(source_path, source);
-        }
-
-        Ok(())
-    }
-
-    /// Load all template files into the database
-    pub fn load_templates(&mut self) -> Result<()> {
-        let templates_dir = self.templates_dir();
-        if !templates_dir.exists() {
-            return Ok(()); // No templates directory is fine
-        }
-
-        let template_files: Vec<Utf8PathBuf> = WalkBuilder::new(&templates_dir)
-            .build()
-            .filter_map(|e| e.ok())
-            .filter(|e| e.file_type().map(|ft| ft.is_file()).unwrap_or(false))
-            .filter(|e| {
-                e.path()
-                    .extension()
-                    .map(|ext| ext == "html")
-                    .unwrap_or(false)
-            })
-            .filter_map(|e| Utf8PathBuf::from_path_buf(e.into_path()).ok())
-            .collect();
-
-        for path in template_files {
-            let content = fs::read_to_string(&path)?;
-            let relative = path
-                .strip_prefix(&templates_dir)
-                .map(|p| p.to_string())
-                .unwrap_or_else(|_| path.to_string());
-
-            let template_path = TemplatePath::new(relative);
-            let template_content = TemplateContent::new(content);
-            let template = TemplateFile::new(&*self.db, template_path.clone(), template_content)?;
-            self.templates.insert(template_path, template);
-        }
-
-        Ok(())
-    }
-
-    /// Load all Sass/SCSS files into the database
-    pub fn load_sass(&mut self) -> Result<()> {
-        let sass_dir = self.sass_dir();
-        if !sass_dir.exists() {
-            return Ok(()); // No sass directory is fine
-        }
-
-        let sass_files: Vec<Utf8PathBuf> = WalkBuilder::new(&sass_dir)
-            .build()
-            .filter_map(|e| e.ok())
-            .filter(|e| e.file_type().map(|ft| ft.is_file()).unwrap_or(false))
-            .filter(|e| {
-                e.path()
-                    .extension()
-                    .map(|ext| ext == "scss" || ext == "sass")
-                    .unwrap_or(false)
-            })
-            .filter_map(|e| Utf8PathBuf::from_path_buf(e.into_path()).ok())
-            .collect();
-
-        for path in sass_files {
-            let content = fs::read_to_string(&path)?;
-            let relative = path
-                .strip_prefix(&sass_dir)
-                .map(|p| p.to_string())
-                .unwrap_or_else(|_| path.to_string());
-
-            let sass_path = SassPath::new(relative);
-            let sass_content = SassContent::new(content);
-            let sass_file = SassFile::new(&*self.db, sass_path.clone(), sass_content)?;
-            self.sass_files.insert(sass_path, sass_file);
-        }
-
-        Ok(())
-    }
-
-    /// Load all static files into the database (from static/ and dist/, with dist/ taking priority)
-    pub fn load_static(&mut self) -> Result<()> {
-        let static_dir = self.static_dir();
-        let dist_dir = self.dist_dir();
-
-        // Load from static/ first
-        if static_dir.exists() {
-            let static_files: Vec<Utf8PathBuf> = WalkBuilder::new(&static_dir)
-                .build()
-                .filter_map(|e| e.ok())
-                .filter(|e| {
-                    e.file_type()
-                        .map(|ft| ft.is_file() || (ft.is_symlink() && e.path().is_file()))
-                        .unwrap_or(false)
-                })
-                .filter_map(|e| Utf8PathBuf::from_path_buf(e.into_path()).ok())
-                .collect();
-
-            for path in static_files {
-                let content = fs::read(&path)?;
-                let relative = path
-                    .strip_prefix(&static_dir)
-                    .map(|p| p.to_string())
-                    .unwrap_or_else(|_| path.to_string());
-
-                let static_path = StaticPath::new(relative);
-                let static_file = StaticFile::new(&*self.db, static_path.clone(), content)?;
-                self.static_files.insert(static_path, static_file);
-            }
-        }
-
-        // Load from dist/ (overwrites static/ on conflict)
-        if dist_dir.exists() {
-            let dist_files: Vec<Utf8PathBuf> = WalkBuilder::new(&dist_dir)
-                .build()
-                .filter_map(|e| e.ok())
-                .filter(|e| {
-                    e.file_type()
-                        .map(|ft| ft.is_file() || (ft.is_symlink() && e.path().is_file()))
-                        .unwrap_or(false)
-                })
-                .filter_map(|e| Utf8PathBuf::from_path_buf(e.into_path()).ok())
-                .collect();
-
-            for path in dist_files {
-                let content = fs::read(&path)?;
-                let relative = path
-                    .strip_prefix(&dist_dir)
-                    .map(|p| p.to_string())
-                    .unwrap_or_else(|_| path.to_string());
-
-                tracing::trace!(path = %relative, "load_static: loading file from dist");
-
-                let static_path = StaticPath::new(relative);
-                let static_file = StaticFile::new(&*self.db, static_path.clone(), content)?;
-                self.static_files.insert(static_path, static_file);
-            }
-
-            // Explicitly load .vite/manifest.json (hidden file not picked up by WalkBuilder)
-            let manifest_path = dist_dir.join(".vite/manifest.json");
-            if manifest_path.exists() {
-                let content = fs::read(&manifest_path)?;
-                tracing::debug!(bytes = content.len(), "loaded vite manifest");
-                let static_path = StaticPath::new(".vite/manifest.json".to_string());
-                let static_file = StaticFile::new(&*self.db, static_path.clone(), content)?;
-                self.static_files.insert(static_path, static_file);
-            }
-        }
-
-        // Check if vite is configured but manifest is missing
-        let project_dir = self.content_dir.parent().unwrap_or(&self.content_dir);
-        if vite::has_vite_config(project_dir.as_std_path()) {
-            let has_manifest = self
-                .static_files
-                .contains_key(&StaticPath::new(".vite/manifest.json".to_string()));
-            if !has_manifest {
-                let dist_dir = self.dist_dir();
-                let manifest_path = dist_dir.join(".vite/manifest.json");
-                return Err(eyre!(
-                    "Vite is configured but manifest not found.\n\n\
-                    Expected manifest at: {}\n\n\
-                    This usually means one of:\n\
-                    1. Vite build hasn't run yet - try `pnpm run build` in {}\n\
-                    2. vite.config.ts is missing `build.manifest: true`\n\
-                    3. vite.config.ts has a different outDir than 'dist'\n\n\
-                    Looked in:\n\
-                    - {}\n",
-                    manifest_path,
-                    project_dir,
-                    manifest_path,
-                ));
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Load all data files into the database
-    pub fn load_data(&mut self) -> Result<()> {
-        let data_dir = self.data_dir();
-        if !data_dir.exists() {
-            return Ok(()); // No data directory is fine
-        }
-
-        let data_files: Vec<Utf8PathBuf> = WalkBuilder::new(&data_dir)
-            .build()
-            .filter_map(|e| e.ok())
-            .filter(|e| e.file_type().map(|ft| ft.is_file()).unwrap_or(false))
-            .filter(|e| {
-                // Only load supported data file formats
-                e.path()
-                    .extension()
-                    .map(|ext| is_data_file_extension(&ext.to_string_lossy()))
-                    .unwrap_or(false)
-            })
-            .filter_map(|e| Utf8PathBuf::from_path_buf(e.into_path()).ok())
-            .collect();
-
-        for path in data_files {
-            let content = fs::read_to_string(&path)?;
-            let relative = path
-                .strip_prefix(&data_dir)
-                .map(|p| p.to_string())
-                .unwrap_or_else(|_| path.to_string());
-
-            let data_path = DataPath::new(relative);
-            let data_content = DataContent::new(content);
-            let data_file = DataFile::new(&*self.db, data_path.clone(), data_content)?;
-            self.data_files.insert(data_path, data_file);
-        }
-
-        Ok(())
-    }
-
-    /// Update a single source file (for incremental rebuilds)
-    pub fn update_source(&mut self, relative_path: &SourcePathRef) -> Result<bool> {
-        let full_path = self.content_dir.join(relative_path.as_str());
-        if !full_path.exists() {
-            // File was deleted
-            self.sources.remove(relative_path);
-            return Ok(true);
-        }
-
-        let content = fs::read_to_string(&full_path)?;
-        let source_content = SourceContent::new(content);
-        let last_modified = fs::metadata(&full_path)?
-            .modified()?
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs() as i64)
-            .unwrap_or(0);
-
-        // Check if we already have this source - for picante, recreate and replace
-        let source_path = SourcePath::new(relative_path.to_string());
-        let source = SourceFile::new(
-            &*self.db,
-            source_path.clone(),
-            source_content,
-            last_modified,
-        )
-        .expect("failed to create source file");
-        self.sources.insert(source_path, source);
-
-        Ok(true)
-    }
-
-    /// Update a single template file (for incremental rebuilds)
-    pub fn update_template(&mut self, relative_path: &TemplatePathRef) -> Result<bool> {
-        let templates_dir = self.templates_dir();
-        let full_path = templates_dir.join(relative_path.as_str());
-        if !full_path.exists() {
-            // File was deleted
-            self.templates.remove(relative_path);
-            return Ok(true);
-        }
-
-        let content = fs::read_to_string(&full_path)?;
-        let template_content = TemplateContent::new(content);
-
-        // For picante, recreate and replace
-        let template_path = TemplatePath::new(relative_path.to_string());
-        let template = TemplateFile::new(&*self.db, template_path.clone(), template_content)
-            .expect("failed to create template file");
-        self.templates.insert(template_path, template);
-
-        Ok(true)
-    }
-
-    /// Update a single Sass file (for incremental rebuilds)
-    pub fn update_sass(&mut self, relative_path: &SassPathRef) -> Result<bool> {
-        let sass_dir = self.sass_dir();
-        let full_path = sass_dir.join(relative_path.as_str());
-        if !full_path.exists() {
-            // File was deleted
-            self.sass_files.remove(relative_path);
-            return Ok(true);
-        }
-
-        let content = fs::read_to_string(&full_path)?;
-        let sass_content = SassContent::new(content);
-
-        // For picante, recreate and replace
-        let sass_path = SassPath::new(relative_path.to_string());
-        let sass_file = SassFile::new(&*self.db, sass_path.clone(), sass_content)
-            .expect("failed to create sass file");
-        self.sass_files.insert(sass_path, sass_file);
-
-        Ok(true)
     }
 }
 
@@ -1297,7 +860,7 @@ pub async fn build(
                 }
                 queries::SiteError::Render(render_error) => {
                     // Use ariadne for pretty ANSI formatting
-                    let formatted = crate::error_pages::format_error_ansi(&render_error.error);
+                    let formatted = dodeca::error_pages::format_error_ansi(&render_error.error);
                     eprintln!(
                         "{} Error rendering {}:\n{}",
                         "✗".red(),
@@ -1598,7 +1161,7 @@ pub async fn build(
                     std::collections::HashMap::new();
                 let mut diagnostics_by_url: std::collections::HashMap<
                     &str,
-                    &crate::db::HttpErrorDiagnostics,
+                    &dodeca::db::HttpErrorDiagnostics,
                 > = std::collections::HashMap::new();
 
                 for broken in &link_result.broken_links {
@@ -2037,13 +1600,13 @@ async fn start_file_watcher(
         }
     });
 
-    crate::spawn::spawn(async move {
+    dodeca::spawn::spawn(async move {
         use tokio::time::{Duration, Instant};
 
         let debounce = Duration::from_millis(100);
         let max_debounce = Duration::from_millis(500);
         let mut pending: Vec<file_watcher::FileEvent> = Vec::new();
-        let mut active_revision: Option<crate::revision::RevisionToken> = None;
+        let mut active_revision: Option<dodeca::revision::RevisionToken> = None;
 
         loop {
             let first = match event_rx.recv().await {
@@ -2317,7 +1880,7 @@ async fn serve_plain(
 
     // Start the cell server in background ASAP so we can accept connections early.
     let server_clone = server.clone();
-    crate::spawn::spawn(async move {
+    dodeca::spawn::spawn(async move {
         if let Err(e) = cell_server::start_cell_server_with_shutdown(
             server_clone,
             cell_path,
@@ -3042,7 +2605,7 @@ async fn serve_with_tui(
                         cas_dir: Utf8PathBuf,
                         code_exec_dir: Utf8PathBuf,
                         cell_path: std::path::PathBuf| {
-        crate::spawn::spawn(async move {
+        dodeca::spawn::spawn(async move {
             let ips = get_bind_ips(mode);
             let requested_port = preferred_port.unwrap_or(4000);
 
@@ -3056,7 +2619,7 @@ async fn serve_with_tui(
             let cell_path_clone = cell_path.clone();
             let event_tx_clone = event_tx.clone();
 
-            let server_task = crate::spawn::spawn(async move {
+            let server_task = dodeca::spawn::spawn(async move {
                 if let Err(e) = cell_server::start_cell_server_with_shutdown(
                     server_clone,
                     cell_path_clone,
@@ -3209,7 +2772,7 @@ async fn serve_with_tui(
     let current_shutdown = Arc::new(std::sync::Mutex::new(shutdown_tx.clone()));
     let current_shutdown_for_handler = current_shutdown.clone();
 
-    crate::spawn::spawn(async move {
+    dodeca::spawn::spawn(async move {
         while let Some(cmd) = cmd_rx.recv().await {
             let new_mode = match cmd {
                 tui::ServerCommand::GoPublic => tui::BindMode::Lan,
@@ -3258,7 +2821,7 @@ async fn serve_with_tui(
     let mut proto_cmd_rx = proto_cmd_rx; // Move the receiver from earlier
     let filter_handle_for_bridge = filter_handle.clone();
     let event_tx_for_bridge = event_tx.clone();
-    crate::spawn::spawn(async move {
+    dodeca::spawn::spawn(async move {
         while let Some(proto_cmd) = proto_cmd_rx.recv().await {
             match proto_cmd {
                 cell_tui_proto::ServerCommand::CycleLogLevel => {
@@ -3322,7 +2885,7 @@ async fn serve_with_tui(
     // Spawn forwarders to push updates to TUI cell via RPC
     // Forward progress updates
     let tui_client_progress = tui_client.clone();
-    crate::spawn::spawn(async move {
+    dodeca::spawn::spawn(async move {
         while progress_rx.changed().await.is_ok() {
             let progress = progress_rx.borrow().clone();
             let _ = tui_client_progress
@@ -3333,7 +2896,7 @@ async fn serve_with_tui(
 
     // Forward server status updates
     let tui_client_status = tui_client.clone();
-    crate::spawn::spawn(async move {
+    dodeca::spawn::spawn(async move {
         while server_rx.changed().await.is_ok() {
             let status = server_rx.borrow().clone();
             let _ = tui_client_status
@@ -3507,7 +3070,7 @@ async fn serve_static(
 
     // Start the cell server with our static content service
     let content_service_clone = content_service.clone();
-    crate::spawn::spawn(async move {
+    dodeca::spawn::spawn(async move {
         if let Err(e) = cell_server::start_static_cell_server(
             content_service_clone,
             cell_path,
