+++ crates/integration-tests/src/tests/mod.rs
@@ -2,6 +2,8 @@ use crate::harness::*;
 use crate::*;
 use std::time::Duration;
 
+#[cfg(test)]
+pub mod authoring_lsp;
 pub mod basic;
 pub mod boot_contract;
 pub mod build_steps;
